refactor: move TryFrom<Map> to object mods
