+++ stac/src/collection.rs
@@ -1,4 +1,4 @@
-use crate::{Asset, Href, Link, Links, STAC_VERSION};
+use crate::{Asset, Error, Href, Link, Links, Result, STAC_VERSION};
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 use std::collections::HashMap;
@@ -239,14 +239,25 @@ impl Default for TemporalExtent {
     }
 }
 
-fn deserialize_type<'de, D>(deserializer: D) -> Result<String, D::Error>
+impl TryFrom<Collection> for Map<String, Value> {
+    type Error = Error;
+    fn try_from(collection: Collection) -> Result<Self> {
+        if let serde_json::Value::Object(object) = serde_json::to_value(collection)? {
+            Ok(object)
+        } else {
+            panic!("all STAC collections should serialize to a serde_json::Value::Object")
+        }
+    }
+}
+
+fn deserialize_type<'de, D>(deserializer: D) -> std::result::Result<String, D::Error>
 where
     D: serde::de::Deserializer<'de>,
 {
     crate::deserialize_type(deserializer, COLLECTION_TYPE)
 }
 
-fn serialize_type<S>(r#type: &String, serializer: S) -> Result<S::Ok, S::Error>
+fn serialize_type<S>(r#type: &String, serializer: S) -> std::result::Result<S::Ok, S::Error>
 where
     S: serde::ser::Serializer,
 {
