+++ stac/src/item.rs
@@ -1,4 +1,4 @@
-use crate::{Asset, Href, Link, Links, STAC_VERSION};
+use crate::{Asset, Error, Href, Link, Links, Result, STAC_VERSION};
 use chrono::Utc;
 use geojson::Geometry;
 use serde::{Deserialize, Serialize};
@@ -183,14 +183,25 @@ impl Links for Item {
     }
 }
 
-fn deserialize_type<'de, D>(deserializer: D) -> Result<String, D::Error>
+impl TryFrom<Item> for Map<String, Value> {
+    type Error = Error;
+    fn try_from(item: Item) -> Result<Self> {
+        if let serde_json::Value::Object(object) = serde_json::to_value(item)? {
+            Ok(object)
+        } else {
+            panic!("all STAC items should serialize to a serde_json::Value::Object")
+        }
+    }
+}
+
+fn deserialize_type<'de, D>(deserializer: D) -> std::result::Result<String, D::Error>
 where
     D: serde::de::Deserializer<'de>,
 {
     crate::deserialize_type(deserializer, ITEM_TYPE)
 }
 
-fn serialize_type<S>(r#type: &String, serializer: S) -> Result<S::Ok, S::Error>
+fn serialize_type<S>(r#type: &String, serializer: S) -> std::result::Result<S::Ok, S::Error>
 where
     S: serde::ser::Serializer,
 {
